Fixed a crash in ConfigGUI when trying to sort by map size and a map file is missing (thanks Moose) (port from 0.61x). Fixed occasional incorrect percentages being printed by /Measure, /Cancel, and /Undo (due to an integer overflow) (ported from 0.61x). Ported recent LineWrapper changes from 0.61x.

+++ branch-0.70x/fCraft.Core/Drawing/DrawOperation.cs
@@ -89,7 +89,7 @@ namespace fCraft.Drawing {
                 }else if( IsDone ) {
                     return 100;
                 } else {
-                    return Math.Min( 100, Math.Max( 0, (BlocksProcessed * 100) / BlocksTotalEstimate ) );
+                    return (int)Math.Min( 100, Math.Max( 0, (BlocksProcessed * 100L) / BlocksTotalEstimate ) );
                 }
             }
         }
